requirements.txt
faiss-cpu
langchain
langchain_community
langchain-openai
python-dotenv
OpenAI
langchain_experimental
pandas
tabulate
google-api-python-client
google-auth-httplib2
google-auth-oauthlib